# 8276662: Scalability bottleneck in SymbolTable::lookup_common()