--- a/src/hotspot/share/classfile/symbolTable.cpp
+++ b/src/hotspot/share/classfile/symbolTable.cpp
@@ -134,6 +134,9 @@ public:
   // We use default allocation/deallocation but counted
   static void* allocate_node(void* context, size_t size, Value const& value) {
     SymbolTable::item_added();
+    if (_lookup_shared_first && (_items_count > 1024)) {
+        _lookup_shared_first = false;
+    }
     return AllocateHeap(size, mtSymbol);
   }
   static void free_node(void* context, void* memory, Value const& value) {
@@ -173,6 +176,12 @@ void SymbolTable::create_table ()  {
   } else {
     _arena = new (mtSymbol) Arena(mtSymbol, symbol_alloc_arena_size);
   }
+  
+#if INCLUDE_CDS
+  _lookup_shared_first = !_shared_table.empty();
+#else
+  _lookup_shared_first = false;
+#endif
 }
 
 void SymbolTable::delete_symbol(Symbol* sym) {
@@ -314,16 +323,12 @@ Symbol* SymbolTable::lookup_common(const char* name,
   if (_lookup_shared_first) {
     sym = lookup_shared(name, len, hash);
     if (sym == NULL) {
-      _lookup_shared_first = false;
       sym = lookup_dynamic(name, len, hash);
     }
   } else {
     sym = lookup_dynamic(name, len, hash);
     if (sym == NULL) {
       sym = lookup_shared(name, len, hash);
-      if (sym != NULL) {
-        _lookup_shared_first = true;
-      }
     }
   }
   return sym;