@shipilev 2021-11-16
I have been looking at this code in JDK-8260718 too. The switching heuristics needs some work, and some experimentation. I think for this sharing fix, slapping `THREAD_LOCAL` to `_lookup_shared_first` would be enough? This would throw away any questions about the changes to switching heuristics. We have one defined already, see e.g. `globalDefinitions_gcc.hpp`:

```
#define THREAD_LOCAL __thread
```