@shipilev 2021-11-16
> (Not that I know what USE_LIBRARY_BASED_TLS_ONLY is really for :-)

Some history: [JDK-8259372](https://bugs.openjdk.java.net/browse/JDK-8259372). Yes, it looks like generic code needs to be wrapped by `USE_LIBRARY_BASED_TLS_ONLY` for a time being. Like it is in `thread.cpp`, if we look for existing code.