@cl4es 2021-11-16
I agree the hint provided by `_lookup_shared_first` makes more sense as a thread local than global state, and that this particular flag seem a worthwhile THREAD_LOCAL candidate in face of potential footprint overhead.