@iklam 2021-11-19
@dwhite-intel do you have any performance results that you can share? Also, the PR description needs to be updated to reflect the final version.