@dwhite-intel 2021-11-19
@iklam, we've seen 5-15% improvement in serialization benchmarks on CascadeLake/IceLake systems, and almost 2x improvement on mutli-socket systems with more cores.